Fix review notes, use more stable way of generating context serializer IDs

+++ rd-net/RdFramework/Reflection/ReflectionSerializers.cs
@@ -670,7 +670,7 @@ namespace JetBrains.Rd.Reflection
         myStore = store;
       }
 
-      public void Register<T>(CtxReadDelegate<T> reader, CtxWriteDelegate<T> writer, int? predefinedType = null)
+      public void Register<T>(CtxReadDelegate<T> reader, CtxWriteDelegate<T> writer, long? predefinedType = null)
       {
         myStore[typeof(T)] = new SerializerPair(reader, writer);
       }
